etl/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# data_maker.py: 1
PyYAML == 3.13

# setup_db.py: 3
# utils.py: 7
SQLAlchemy == 1.2.17

# helpers.py: 1
pandas == 0.24.0

# utils.py: 2
requests == 2.21.0

psycopg2==2.7.7
psycopg2-binary==2.7.7
